refactor kona host to library
