+++ bin/host/src/cli/mod.rs
@@ -14,10 +14,10 @@ mod parser;
 pub(crate) use parser::parse_b256;
 
 mod tracing_util;
-pub(crate) use tracing_util::init_tracing_subscriber;
+pub use tracing_util::init_tracing_subscriber;
 
 /// The host binary CLI application arguments.
-#[derive(Parser, Serialize, Clone)]
+#[derive(Parser, Serialize, Clone, Debug)]
 pub struct HostCli {
     /// Verbosity level (0-4)
     #[arg(long, short, help = "Verbosity level (0-4)", action = ArgAction::Count)]
