+++ bin/host/src/server.rs
@@ -14,6 +14,7 @@ use tokio::sync::RwLock;
 
 /// The [PreimageServer] is responsible for waiting for incoming preimage requests and
 /// serving them to the client.
+#[derive(Debug)]
 pub struct PreimageServer<P, H, KV>
 where
     P: PreimageOracleServer,
